feat(models): refactor models to module, utilize pydantic
