+++ threedframe/models/__init__.py
@@ -1,0 +1,5 @@
+"""3dframe models module."""
+
+from .mesh import MeshData, MeshFace, MeshPoint
+from .joint import JointFixture
+from .model import ModelData, ModelEdge, ModelVertex
