+++ threedframe/utils.py
@@ -3,26 +3,22 @@
 """3DFrame Model Generator Utils."""
 import sys
 import time
-import string
-import itertools
 import subprocess as sp
-from typing import Dict, List, Tuple, Union, Callable, Iterator, Optional, Sequence
+from typing import List, Tuple, Union, Callable, Sequence
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from dataclasses import field, dataclass
 
+import sh
 import numpy as np
 import sympy as S
 from rich import print
 from solid import OpenSCADObject, text, union, resize, translate, scad_render, linear_extrude
-from euclid3 import Point3, Vector3
+from euclid3 import Point3
 from rich.text import Text
 from rich.console import RenderableType
 from rich.progress import Task, TextColumn, SpinnerColumn, ProgressColumn
 from watchdog.events import PatternMatchingEventHandler
 from watchdog.observers import Observer
-
-from threedframe import mesh as meshutil
 
 
 class ComputeTestResultsColumn(ProgressColumn):
@@ -65,243 +61,6 @@ class ParentSpinnerColumn(SpinnerColumn):
         return Text.from_markup("")
 
 
-def label_generator() -> Iterator[str]:
-    """Generates labels for vertices.
-
-    Yields: 'AA', 'AB', 'AC'...'ZW', 'ZY', 'ZZ'
-
-    """
-    base_charmap = iter(string.ascii_uppercase)
-    _label_charmap = iter(string.ascii_uppercase)
-    _base_label = None
-    while True:
-        if not _base_label:
-            _base_label = next(base_charmap)
-        try:
-            label = next(_label_charmap)
-        except StopIteration:
-            try:
-                _base_label = next(base_charmap)
-            except StopIteration:
-                break
-            _label_charmap = iter(string.ascii_uppercase)
-            label = next(_label_charmap)
-        yield f"{_base_label}{label}"
-
-
-MODEL_LABELS = label_generator()
-
-
-@dataclass
-class ModelData:
-    """Computed model info."""
-
-    # Total number of vertices in model.
-    num_vertices: int
-    # Total number of edges in model.
-    num_edges: int
-    # Vertices.
-    vertices: Dict[int, "ModelVertex"]
-
-    def get_edge_target_vertex(self, edge: "ModelEdge") -> "ModelVertex":
-        """Retrieve an edges target vertex."""
-        vertex = self.vertices[edge.target_vidx]
-        return vertex
-
-
-@dataclass
-class ModelEdge:
-    """Computed edge info."""
-
-    # Edge index.
-    eidx: int
-    # Edge length.
-    length: float
-    # Joint vertex index.
-    joint_vidx: int
-    # Target vertex index.
-    target_vidx: int
-    # Vector FROM target vertex into joint.
-    vector_ingress: Tuple[float, float, float]
-
-    @property
-    def length_in(self) -> float:
-        """Length in inches."""
-        return self.length / 25.4
-
-
-@dataclass
-class ModelVertex:
-    """Computed Vertex info."""
-
-    # Vertex index.
-    vidx: int
-    # Edge map.
-    edges: List[ModelEdge]
-    # Vertex Point.
-    point: Tuple[float, float, float]
-    # Vertex Point normal.
-    point_normal: Tuple[float, float, float]
-    # Generated label for vertex.
-    label: str = field(default_factory=lambda: next(MODEL_LABELS))
-
-
-@dataclass
-class MeshPoint:
-    """Mesh Point."""
-
-    # Vertex Index.
-    vidx: int
-    # 3D Coordinate of point.
-    point: Tuple[float, float, float]
-    # Point normal.
-    normal: Optional[Tuple[float, float, float]]
-    # Parent Mesh Data.
-    _parent: Optional["MeshData"] = field(repr=False, hash=False, default=None)
-
-    @property
-    def faces(self) -> List["MeshFace"]:
-        return [f for f in self._parent.faces if self.vidx in f.vertex_indices]
-
-    @property
-    def as_euclid(self) -> Point3:
-        return Point3(*self.point)
-
-    @property
-    def as_sympy(self) -> S.Point3D:
-        return S.Point3D(*self.point)
-
-    @property
-    def as_vector(self) -> Vector3:
-        return Vector3(*self.point)
-
-    @property
-    def normal_vector(self) -> Vector3:
-        return Vector3(*self.normal)
-
-
-@dataclass
-class MeshFace:
-    """Mesh Face."""
-
-    # Face Index.
-    fidx: int
-    # Vertices that makeup face.
-    vertex_indices: List[int]
-    # Face normal.
-    normal: Optional[Tuple[float, float, float]]
-    # Face area.
-    area: float
-    # Vector representing face centroids.
-    centroid: Optional[Tuple[float, float, float]]
-    # Parent Mesh Data.
-    _parent: Optional["MeshData"] = field(repr=False, hash=False, default=None)
-
-    @property
-    def vertices(self) -> List[MeshPoint]:
-        return [v for v in self._parent.vertices if v.vidx in self.vertex_indices]
-
-    @property
-    def sympy_vertices(self) -> List[S.Point3D]:
-        return [p.as_sympy for p in self.vertices]
-
-    @property
-    def euclid_vertices(self) -> List[Point3]:
-        return [p.as_euclid for p in self.vertices]
-
-    @property
-    def as_sympy_plane(self) -> S.Plane:
-        return S.Plane(*self.sympy_vertices, normal_vector=self.normal)
-
-    @property
-    def missing_rect_vertex(self) -> S.Point3D:
-        points_perms = itertools.permutations(self.sympy_vertices, 3)
-        least_dist = None
-        least_dist_point = None
-        for perm in points_perms:
-            midp = find_missing_rect_vertex(*perm)
-            dist = self.centroid_point.distance(midp)
-            if least_dist is None or least_dist > dist:
-                least_dist = dist
-                least_dist_point = midp
-        return least_dist_point
-
-    @property
-    def midpoint_by_canberra(self) -> S.Point3D:
-        fp_1 = self.vertices[0].as_sympy
-        fp_2 = max(self.sympy_vertices, key=lambda p: fp_1.canberra_distance(p))
-        return fp_1.midpoint(fp_2)
-
-    @property
-    def normal_vector(self) -> Vector3:
-        return Vector3(*self.normal)
-
-    @property
-    def centroid_point(self) -> S.Point3D:
-        return S.Point3D(*self.centroid)
-
-    # def calc_absolute_midpoint(self):
-    #     return find_center_of_gravity(*self.sympy_vertices)
-
-
-@dataclass
-class MeshData:
-    """Mesh Analysis Data."""
-
-    # Vertices of mesh.
-    vertices: List[MeshPoint]
-    # Faces of mesh.
-    faces: List[MeshFace]
-
-    def __post_init__(self):
-        for vert in self.vertices:
-            vert._parent = self
-        for face in self.faces:
-            face._parent = self
-
-    @classmethod
-    def from_dict(cls, data):
-        vertices = data.get("vertices", [])
-        faces = data.get("faces", [])
-        verts = [MeshPoint(**v) for v in vertices]
-        fces = [MeshFace(**f) for f in faces]
-        return cls(vertices=verts, faces=fces)
-
-    def calc_absolute_midpoint(self):
-        face_mps = []
-        for face in self.faces:
-            face_mps.append(find_center_of_gravity(*face.sympy_vertices, face.missing_rect_vertex))
-        return find_center_of_gravity(*face_mps)
-
-
-@dataclass
-class JointFixture:
-    """Joint Fixture item."""
-
-    # OpenSCAD Object.
-    scad_object: OpenSCADObject
-    # OpenSCAD inspection object (no hollow).
-    inspect_object: OpenSCADObject
-    # Inner fixture Object (the inner shell).
-    inner_object: OpenSCADObject
-    # Edge item of fixture.
-    model_edge: ModelEdge
-    # Vertex item of fixture.
-    model_vertex: ModelVertex
-    # Destination point of fixture.
-    dest_point: Tuple[float, float, float]
-    # Destination normal of fixture.
-    dest_normal: Tuple[float, float, float]
-    # Inspect data (deprecated)
-    inspect_data: Dict
-    # Mesh Data of inspect object.
-    inspect_mesh: MeshData = None
-
-    def __post_init__(self):
-        if not self.inspect_mesh:
-            self.inspect_mesh = analyze_scad(self.inspect_object)
-
-
 def round_point(point: Point3, n_digits=2):
     """Round a given 3d point."""
     p = point.copy()
@@ -320,8 +79,9 @@ def openscad_cmd(*args) -> sp.Popen:
 
 def exec_blender_script(model_path: Path, script_path: Path, out_path: Path):
     """Execute blender command."""
+    blender_loc = sh.which("blender")
     _cmd = [
-        "/usr/local/blender/blender",
+        blender_loc,
         str(model_path.absolute()),
         "--background",
         "--python-use-system-env",
@@ -371,14 +131,6 @@ class TemporaryScadWorkspace(TemporaryDirectory):
                 self._stream_proc(self.proc)
             files.append((name, obj_path, render_path))
         return self.path, files
-
-
-def analyze_scad(obj: OpenSCADObject) -> MeshData:
-    scad_obj = [("target", union()(obj), "stl")]
-    with TemporaryScadWorkspace(scad_objs=scad_obj) as tmpdata:
-        tmp_path, tmp_files = tmpdata
-        data = meshutil.analyze_mesh(tmp_files[0][-1])
-        return MeshData.from_dict(data)
 
 
 # Fixed version of label size.
